Refactor logic for pipeline and n_serializer
Both abstractions use consumer_bounded_queue as an underlying
implementation. This is now moved in its own header in details/
