+++ src/n_serializer.cpp
@@ -1,7 +1,7 @@
 #include "concore/n_serializer.hpp"
 #include "concore/global_executor.hpp"
 #include "concore/spawn.hpp"
-#include "concore/data/concurrent_queue.hpp"
+#include "concore/detail/consumer_bounded_queue.hpp"
 #include "concore/detail/utils.hpp"
 #include "concore/detail/enqueue_next.hpp"
 
@@ -17,27 +17,15 @@ struct n_serializer::impl : std::enable_shared_from_this<impl> {
     executor_t base_executor_;
     //! The executor to be used when
     executor_t cont_executor_;
-    //! The number of tasks that can be executed in parallel
-    int max_par_;
     //! Handler to be called whenever we have an exception while enqueueing the next task
     except_fun_t except_fun_;
-    //! The queue of tasks that wait to be executed
-    concurrent_queue<task, queue_type::multi_prod_multi_cons> waiting_tasks_;
-    //! The number of tasks that are in the queue, active or not
-    std::atomic<uint32_t> combined_count_{0};
-
-    union count_bits {
-        uint32_t int_value;
-        struct {
-            unsigned num_active : 16;
-            unsigned num_tasks : 16;
-        } fields;
-    };
+    //! The tasks that are enqueued into our object
+    detail::consumer_bounded_queue<task> processing_items_;
 
     impl(int N, executor_t base_executor, executor_t cont_executor)
         : base_executor_(base_executor)
         , cont_executor_(cont_executor)
-        , max_par_(N) {
+        , processing_items_(N) {
         if (!base_executor)
             base_executor_ = global_executor;
         if (!cont_executor)
@@ -45,22 +33,7 @@ struct n_serializer::impl : std::enable_shared_from_this<impl> {
     }
 
     void enqueue(task&& t) {
-        // Add the task to the queue
-        waiting_tasks_.push(std::move(t));
-
-        // Increase the number of tasks we keep track of.
-        // Check if we can increase the "active" tasks; if so, enqueue a new task in the base
-        // executor
-        count_bits old, desired;
-        old.int_value = combined_count_.load();
-        do {
-            desired.int_value = old.int_value;
-            desired.fields.num_tasks++;
-            if (desired.fields.num_active < max_par_)
-                desired.fields.num_active++;
-        } while (!combined_count_.compare_exchange_weak(old.int_value, desired.int_value));
-
-        if (desired.fields.num_active != old.fields.num_active)
+        if (processing_items_.push_and_try_acquire(std::move(t)))
             enqueue_next(base_executor_);
     }
 
@@ -70,20 +43,12 @@ struct n_serializer::impl : std::enable_shared_from_this<impl> {
     //! Note: we only execute one task, even if we have multiple tasks. We do this to ensure that
     //! the tasks are relatively small sized.
     void execute_one() {
-        detail::pop_and_execute(waiting_tasks_);
+        // Execute one task
+        auto to_execute = processing_items_.extract_one();
+        detail::execute_task(to_execute);
 
-        // Decrement the number of tasks
-        // Check if we need to enqueue a continuation
-        count_bits old, desired;
-        old.int_value = combined_count_.load();
-        do {
-            desired.int_value = old.int_value;
-            desired.fields.num_tasks--;
-            desired.fields.num_active =
-                    std::min(desired.fields.num_active, desired.fields.num_tasks);
-        } while (!combined_count_.compare_exchange_weak(old.int_value, desired.int_value));
-
-        if (desired.fields.num_active == old.fields.num_active)
+        // Can we start a new task?
+        if (processing_items_.release_and_acquire())
             enqueue_next(cont_executor_);
     }
